Factory: moved Factory pattern to new structure of directories with autoload
