+++ index.php
@@ -1,6 +1,9 @@
 <?php
 require 'vendor/autoload.php';
-use \Patterns\CreationalDesign\{Singleton\Singleton};
+use \Patterns\CreationalDesign\{
+  Singleton\Singleton,
+  Factory\ShapeFactory,
+};
 
 //SINGLETON-------------------------------------------------------------------------------------------------------------
 class Car extends Singleton {
@@ -26,6 +29,18 @@ $volvo = Volvo::getInstance();
 $volvo->mileage = 331;
 
 echo '<pre>';
-echo '<h1>Singleton</h1>';
+echo '<h2>Singleton</h2>';
 var_dump($car, $fiat, $volvo, Car::getInstance(), Fiat::getInstance(), Volvo::getInstance());
 //SINGLETON-END---------------------------------------------------------------------------------------------------------
+
+//FACTORY---------------------------------------------------------------------------------------------------------------
+$shapeFactory = new ShapeFactory();
+
+$circle = $shapeFactory->getShape(ShapeFactory::$SHAPE_CIRCLE);
+$square = $shapeFactory->getShape(ShapeFactory::$SHAPE_SQUARE);
+
+echo '<h2>FACTORY</h2>';
+
+$circle->draw();
+$square->draw();
+//FACTORY-END-----------------------------------------------------------------------------------------------------------
