+++ src/CreationalDesign/Factory/Circle.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Patterns\CreationalDesign\Factory;
+
+class Circle implements Shape
+{
+  public function draw(): void
+  {
+    echo 'Circle' . PHP_EOL;
+  }
+}
