+++ src/CreationalDesign/Factory/Shape.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Patterns\CreationalDesign\Factory;
+
+interface Shape
+{
+  public function draw(): void;
+}
