+++ src/CreationalDesign/Factory/ShapeFactory.php
@@ -1,8 +1,6 @@
 <?php
 
-namespace Factory;
-
-include "Shape.php";
+namespace Patterns\CreationalDesign\Factory;
 
 /**
  * Class ShapeFactory Noob implementation of getShape method (predefined list of possible objects to create)
@@ -13,16 +11,16 @@ class ShapeFactory
   public static $SHAPE_CIRCLE = 'Circle';
   public static $SHAPE_SQUARE = 'Square';
 
-  public function getShape(String $shapeId): Shape\Shape
+  public function getShape(String $shapeId): Shape
   {
     if ($shapeId == null) {
       throw new Exception('Shape can\'t be null.');
     }
 
     if (strcmp($shapeId, 'Circle') == 0) {
-      return new Shape\Circle();
+      return new Circle();
     } else if (strcmp($shapeId, 'Square') == 0) {
-      return new Shape\Square();
+      return new Square();
     }
   }
 }
