+++ src/CreationalDesign/Factory/Square.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Patterns\CreationalDesign\Factory;
+
+
+class Square implements Shape
+{
+  public function draw(): void
+  {
+    echo 'Square' . PHP_EOL;
+  }
+}
